\name{EawagSchoolTools-package}

\alias{EawagSchoolTools-package}
\alias{EawagSchoolTools}

\title{Miscellaneous Tools for Eawag Summer School in Environmental Systems Analysis}

\description{
This package contains a collection of miscellaneous functions for use in the exercises for the Eawag Summer School in Environmental Systems Analysis. Additional functions for the exercises are found in the \code{GASPbias} and \code{SABC} packages.
}

\seealso{
Packages \code{GASPbias} and \code{SABC}.
}

\references{
Reichert, P. (2012). \emph{Environmental Systems Analysis.} Course notes for 6th Eawag Summer School in Environmental Systems Analysis 2014.
}

\author{
Peter Reichert <peter.reichert@eawag.ch>. Package initially compiled by Lukas M. Weber.
}