\name{resid.diag}

\alias{resid.diag}

\title{Plotting functions (eawagSummerSchoolTools)}

\description{
Plot residual diagnostic plots.
}

\usage{
resid.diag(obs,calc,header="")
}

\arguments{
  \item{obs}{vector of observations}
  \item{calc}{vector of calculated results}
  \item{header}{optional header}
}

\value{
Diagnostic plots.
}

\seealso{
\code{\link{resid.diag.boxcox}}
}

\author{
Peter Reichert <peter.reichert@eawag.ch>
}